+++ packages/wayfinder-core/src/routing/round-robin.ts
@@ -15,22 +15,6 @@
  * limitations under the License.
  */
 import { defaultLogger } from '../logger.js';
-/**
- * WayFinder
- * Copyright (C) 2022-2025 Permanent Data Solutions, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 import type { Logger, RoutingStrategy } from '../types.js';
 
 export class RoundRobinRoutingStrategy implements RoutingStrategy {
@@ -51,24 +35,9 @@ export class RoundRobinRoutingStrategy implements RoutingStrategy {
     this.logger = logger;
   }
 
-  // provided gateways are ignored
-  async selectGateway({
-    gateways = [],
-  }: {
-    gateways?: URL[];
-    path?: string;
-    subdomain?: string;
-  } = {}): Promise<URL> {
-    if (gateways.length > 0) {
-      this.logger.warn(
-        'RoundRobinRoutingStrategy does not accept provided gateways. Ignoring provided gateways...',
-        {
-          providedGateways: gateways.length,
-          internalGateways: this.gateways,
-        },
-      );
-    }
+  async selectGateway(): Promise<URL> {
     const gateway = this.gateways[this.currentIndex];
+    this.logger.info('Selecting gateway', { gateway });
     this.currentIndex = (this.currentIndex + 1) % this.gateways.length;
     return gateway;
   }