+++ packages/wayfinder-core/src/types.ts
@@ -187,12 +187,12 @@ export interface TelemetrySettings {
 export type DataStream = ReadableStream<Uint8Array> | AsyncIterable<Uint8Array>;
 
 export interface GatewaysProvider {
-  getGateways(params?: { path?: string; subdomain?: string }): Promise<URL[]>;
+  getGateways(): Promise<URL[]>;
 }
 
 export interface RoutingStrategy {
   selectGateway(params: {
-    gateways: URL[];
+    gateways?: URL[];
     path?: string;
     subdomain?: string;
   }): Promise<URL>;
@@ -200,7 +200,11 @@ export interface RoutingStrategy {
 
 export interface VerificationStrategy {
   trustedGateways: URL[];
-  verifyData(params: { data: DataStream; txId: string }): Promise<void>;
+  verifyData(params: {
+    data: DataStream;
+    headers: Record<string, string>;
+    txId: string;
+  }): Promise<void>;
 }
 
 export interface DataClassifier {